Fix emulator-28.0.5 by using emulator-canary. Add x86_64 images
The bump from 28.0.3 to 29.0.5 broke emulators on travis, but happily
just moving to the canary fixed the existing emulators and added the ability
to run x86_64 images unaccelerated, so we have modern APIs.

Still work to do (relaxing timings mostly) to get 28+Q working but it's possible.

Upstream issue + discussion here
ttps://travis-ci.community/t/android-emulators-not-starting-for-the-last-few-days-late-march-2019/2871/4