Preference: Disable "Card Browser" context menu
We can do this via defining an activity alias, then disabling it in the
system.

According to StackOverflow, this is persisted until uninstall/reinstall
but it's unknown whether this affects updates.

It doesn't matter, as we perform the check on startup.

This commit also changes the menu to make it localizable. This is based
on the system language.

Fixes #6226