Manually bump version after failed experiment
Experiment: what happens if you merge a PR while the Publish Action is working?

You end up with git rejecting the Publish Action's version bump commit because it
is behind HEAD on the main branch.

Follow-on task is to modify the action to definitely tag the commit that was built,
but to fast-forward to current HEAD before pushing so that it self-heals in these cases
to a consistent version number on HEAD of main branch